BBVAFSW3T3-26 | Refactor de UserService
