+++ wallet/src/main/java/com/bbva/wallet/services/impl/UserServiceImpl.java
@@ -1,0 +1,22 @@
+package com.bbva.wallet.services.impl;
+
+import com.bbva.wallet.repositories.UserRepository;
+import com.bbva.wallet.services.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public void deleteById(Long id) {
+        var user = userRepository.findById(id);
+        user.ifPresent(value -> {
+            value.setSoftDelete(true);
+            userRepository.save(value);
+        });
+    }
+}
